+++ src/js/modules/helpers/transparency.js
